pyproject.toml
[tool.poetry]
name = "bubla"
version = "0.1.0"
description = "personal discord bot for learning discord api and stuff"
authors = ["Zdenek Kraus <zdenek.kraus@gmail.com>"]
readme = "README.md"
packages = [{ include = "bubla" }]

[tool.poetry.dependencies]
python = ">3.10"
discord-py = "^2.3.2"
colorlog = "^6.7.0"
markdownify = "^0.11.6"
google-api-python-client = "^2.108.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
google = "^3.0.0"
requests = "^2.31.0"
tabulate = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = ["poetry.core.masonry.api"]